Add getValuesIncomeAndExpense util function to Transfer component
